# backend/src/routes/leaderboard.js
/**
 * src/routes/leaderboard.js
 */
"use strict";
const express = require("express");
const router = express.Router();
const pool = require("../db/pool");
const redis = require("../services/redis");
const { leaderboardQueryDuration } = require("../services/metrics");
const { createRateLimiter } = require("../middleware/rateLimiter");

// 30 requests per minute per IP — prevents enumeration / data scraping (issue #695)
const leaderboardLimiter = createRateLimiter(30, 1, "leaderboard");

// Cursor-based pagination constants (matching /api/projects conventions).
const LEADERBOARD_DEFAULT_LIMIT = 50;
const LEADERBOARD_MAX_LIMIT = 200;

// The ranking query aggregates every donation per profile, so it is far too
// expensive to run per request (issue #1093). 60 seconds is short enough that a
// freshly recorded donation surfaces almost immediately and long enough to
// absorb a burst of reads.
const LEADERBOARD_CACHE_TTL = 60; // seconds
const LEADERBOARD_CACHE_PREFIX = "leaderboard:";

/**
 * Cache key for one leaderboard page.
 *
 * Every parameter that changes the result set is part of the key — including
 * the pagination ones, so a cached page can never be served for a different
 * page or ordering.
 *
 * @param {{ pageSize: number, cursor?: string, period: string, sortBy: string, onlyVerified: boolean }} params
 * @returns {string}
 */
function getLeaderboardCacheKey({ pageSize, cursor, period, sortBy, onlyVerified }) {
  return [
    LEADERBOARD_CACHE_PREFIX,
    sortBy,
    period,
    onlyVerified ? "verified" : "all",
    pageSize,
    cursor || "first",
  ].join(":");
}

router.get("/", leaderboardLimiter, async (req, res, next) => {
  try {
    const pageSize = Math.min(
      Number.parseInt(req.query.limit, 10) || LEADERBOARD_DEFAULT_LIMIT,
      LEADERBOARD_MAX_LIMIT
    );
    const cursor = req.query.cursor;
    const period = req.query.period || "all";
    const sortBy = req.query.sortBy === "impactScore" ? "impact_score" : "total_donated_xlm";
    const onlyVerified = req.query.onlyVerified === "true";

    const conditions = [];
    const params = [];

    if (period === "week") {
      conditions.push("d.created_at >= NOW() - INTERVAL '7 days'");
    } else if (period === "month") {
      conditions.push("d.created_at >= NOW() - INTERVAL '30 days'");
    } else if (period === "year") {
      conditions.push("d.created_at >= NOW() - INTERVAL '1 year'");
    }

    if (onlyVerified) {
      // Distinct aliases avoid clashing with the outer `pr` JOIN used for impact/CO2.
      const verifiedSubQuery = `
        NOT EXISTS (
          SELECT 1 FROM donations d2
          JOIN projects pr_unverified ON d2.project_id = pr_unverified.id
          WHERE d2.donor_address = p.public_key AND pr_unverified.verified = false
        )
        AND EXISTS (
          SELECT 1 FROM donations d3
          JOIN projects pr_verified ON d3.project_id = pr_verified.id
          WHERE d3.donor_address = p.public_key AND pr_verified.verified = true
        )
      `;
      conditions.push(`(${verifiedSubQuery})`);
    }

    // Cursor-based pagination on (sortBy, public_key), mirroring /api/projects.
    if (cursor) {
      let cursorData;
      try {
        cursorData = JSON.parse(Buffer.from(cursor, "base64").toString("utf8"));
      } catch {
        return res.status(400).json({ error: "Invalid cursor" });
      }
      const sortValue = sortBy === "total_donated_xlm"
        ? cursorData.total_donated_xlm
        : sortBy === "total_co2_offset_kg"
          ? cursorData.total_co2_offset_kg
          : cursorData.donor_count;
      // eslint-disable-next-line security/detect-object-injection
      const sortValue = cursorData[sortBy];
      const publicKey = cursorData.publicKey;
      if (sortValue === undefined || !publicKey) {
        return res.status(400).json({ error: "Invalid cursor" });
      }
      params.push(sortValue, publicKey);
      const sortIdx = params.length - 1;
      const keyIdx = params.length;
      conditions.push(
        `(${sortBy} < $${sortIdx} OR (${sortBy} = $${sortIdx} AND p.public_key < $${keyIdx}))`,
      );
    }

    // Fetch pageSize + 1 so we can detect whether another page exists.
    params.push(pageSize + 1);
    const limitIdx = params.length;

    const cacheKey = getLeaderboardCacheKey({ pageSize, cursor, period, sortBy, onlyVerified });
    // A Redis outage makes get() resolve to null, so the endpoint degrades to
    // hitting Postgres rather than failing.
    const cached = await redis.get(cacheKey);
    if (cached) {
      return res.json(cached);
    }

    const whereClause = conditions.length > 0
      ? `WHERE ${conditions.join("\n  AND ")}`
      : "";

    const query = `
      SELECT p.public_key, p.display_name, p.badges,
             COALESCE(SUM(d.amount_xlm), 0)::NUMERIC AS total_donated_xlm,
             COUNT(DISTINCT d.project_id)::INTEGER AS projects_supported,
             COALESCE(
               SUM(
                 CASE
                   WHEN pr.raised_xlm > 0 THEN (d.amount_xlm * (pr.co2_offset_kg::numeric / pr.raised_xlm))
                   ELSE 0
                 END
               ),
               0
             )::NUMERIC AS total_co2_offset_kg,
             (
               COALESCE(SUM(d.amount_xlm), 0) * 0.7 +
               (
                 COALESCE(
                   SUM(
                     CASE
                       WHEN pr.raised_xlm > 0 THEN (d.amount_xlm * (pr.co2_offset_kg::numeric / pr.raised_xlm))
                       ELSE 0
                     END
                   ),
                   0
                 ) / 100
               ) * 0.3
             )::NUMERIC AS impact_score
      FROM profiles p
      LEFT JOIN donations d ON p.public_key = d.donor_address
      LEFT JOIN projects pr ON pr.id = d.project_id
      ${whereClause}
      GROUP BY p.public_key, p.display_name, p.badges
      ORDER BY ${sortBy} DESC, p.public_key DESC
      LIMIT $${limitIdx}
    `;

    // eslint-disable-next-line sql-injection/no-sql-injection
    const observeQueryDuration = leaderboardQueryDuration.startTimer({ period, sort_by: sortBy });
    let result;
    try {
      result = await pool.query(query, params);
    } finally {
      // Recorded even when the query throws, otherwise the failures that matter
      // most would be missing from the histogram.
      observeQueryDuration();
    }
    const rows = result.rows;
    const hasMore = rows.length > pageSize;
    const pageRows = rows.slice(0, pageSize);

    const entries = pageRows.map((p, i) => ({
      rank: i + 1,
      publicKey: p.public_key,
      displayName: p.display_name || null,
      totalDonatedXLM: p.total_donated_xlm?.toString() || "0",
      projectsSupported: p.projects_supported,
      topBadge: p.badges?.[0]?.tier || null,
      impactScore: p.impact_score?.toString() || "0",
      totalCO2OffsetKg: p.total_co2_offset_kg?.toString() || "0",
    }));

    let nextCursor = null;
    if (hasMore) {
      const last = pageRows[pageRows.length - 1];
      const lastSortVal = sortBy === "total_donated_xlm"
        ? last.total_donated_xlm
        : sortBy === "total_co2_offset_kg"
          ? last.total_co2_offset_kg
          : last.donor_count;
      nextCursor = Buffer.from(
        JSON.stringify({ [sortBy]: lastSortVal, publicKey: last.public_key }),
      // eslint-disable-next-line security/detect-object-injection
      nextCursor = Buffer.from(
        // eslint-disable-next-line security/detect-object-injection
        JSON.stringify({ [sortBy]: last[sortBy], publicKey: last.public_key }),
      ).toString("base64");
    }

    const payload = {
      success: true,
      data: entries,
      has_more: hasMore,
      next_cursor: nextCursor,
    };

    // Cache the fully-shaped response: `rank` is positional and derived from
    // `impact_score`/`total_donated_xlm` ordering, so caching the rows alone
    // would mean recomputing — and re-reading — the same page twice.
    await redis.set(cacheKey, payload, LEADERBOARD_CACHE_TTL);

    res.json(payload);
  } catch (e) {
    next(e);
  }
});

/**
 * GET /api/leaderboard/teams
 * Teams ranked by combined total_donated_xlm across their members — the
 * team-giving counterpart to the donor leaderboard. A wallet's donations
 * count toward exactly one team (enforced by a unique index on
 * team_members.wallet_address), so combined totals never double-count.
 *
 * Query params:
 *   - limit (default 50, max 200)
 *   - offset (default 0)
 *   - period (week | month | year | all; default all)
 */
router.get("/teams", leaderboardLimiter, async (req, res, next) => {
  try {
    const limit = Math.min(
      Math.max(Number.parseInt(req.query.limit, 10) || 50, 1),
      200,
    );
    const offset = Math.max(Number.parseInt(req.query.offset, 10) || 0, 0);
    const period = req.query.period || "all";

    let donationWindow = "";
    if (period === "week") donationWindow = "AND d.created_at >= NOW() - INTERVAL '7 days'";
    else if (period === "month") donationWindow = "AND d.created_at >= NOW() - INTERVAL '30 days'";
    else if (period === "year") donationWindow = "AND d.created_at >= NOW() - INTERVAL '1 year'";

    const result = await pool.query(
      `SELECT t.id, t.name, t.logo_url,
              COUNT(DISTINCT tm.wallet_address)::int AS member_count,
              COALESCE(SUM(d.amount_xlm), 0)::NUMERIC AS total_donated_xlm,
              COALESCE(
                SUM(
                  CASE
                    WHEN pr.raised_xlm > 0 THEN (d.amount_xlm * (pr.co2_offset_kg::numeric / pr.raised_xlm))
                    ELSE 0
                  END
                ),
                0
              )::NUMERIC AS total_co2_offset_kg
       FROM teams t
       JOIN team_members tm ON tm.team_id = t.id
       LEFT JOIN donations d ON d.donor_address = tm.wallet_address ${donationWindow}
       LEFT JOIN projects pr ON pr.id = d.project_id
       GROUP BY t.id, t.name, t.logo_url
       HAVING COALESCE(SUM(d.amount_xlm), 0) > 0
       ORDER BY total_donated_xlm DESC, t.id DESC
       LIMIT $1 OFFSET $2`,
      [limit + 1, offset],
    );

    const rows = result.rows;
    const hasMore = rows.length > limit;
    const pageRows = rows.slice(0, limit);

    const entries = pageRows.map((row, i) => ({
      rank: offset + i + 1,
      id: row.id,
      name: row.name,
      logoUrl: row.logo_url || null,
      memberCount: row.member_count,
      totalDonatedXLM: row.total_donated_xlm?.toString() || "0",
      totalCO2OffsetKg: row.total_co2_offset_kg?.toString() || "0",
    }));

    res.json({
      success: true,
      data: entries,
      has_more: hasMore,
      next_offset: hasMore ? offset + limit : null,
    });
  } catch (e) {
    next(e);
  }
});

/**
 * GET /api/leaderboard/history
 * Returns the monthly leaderboard snapshots, grouped by month descending.
 * Query params:
 *   - months (int, max 24, default 12): how many past months to return
 */
router.get("/history", leaderboardLimiter, async (req, res, next) => {
  try {
    const months = Math.min(parseInt(req.query.months, 10) || 12, 24);
    const result = await pool.query(
      `SELECT month, donor_address, display_name, total_xlm_that_month, badge, rank
       FROM monthly_leaderboard
       WHERE month >= DATE_TRUNC('month', NOW()) - ($1 - 1) * INTERVAL '1 month'
       ORDER BY month DESC, rank ASC`,
      [months]
    );

    // Group rows by month using a Map to avoid object injection
    const grouped = new Map();
    for (const row of result.rows) {
      const key = row.month.toISOString().slice(0, 7); // "YYYY-MM"
      if (!grouped.has(key)) grouped.set(key, []);
      grouped.get(key).push({
      // eslint-disable-next-line security/detect-object-injection
      if (!grouped[key]) grouped[key] = [];
      // eslint-disable-next-line security/detect-object-injection
      grouped[key].push({
        rank: row.rank,
        donorAddress: row.donor_address,
        displayName: row.display_name || null,
        totalXLMThatMonth: row.total_xlm_that_month?.toString() || "0",
        badge: row.badge || null,
      });
    }

    const history = Array.from(grouped.entries()).map(([month, entries]) => ({ month, entries }));
    res.json({ success: true, data: history });
  } catch (e) {
    next(e);
  }
});

/**
 * POST /api/leaderboard/snapshot
 * Admin endpoint: snapshot the current month's top donors into monthly_leaderboard.
 * Idempotent — re-running for the same month overwrites existing rows via ON CONFLICT.
 * Requires header: x-admin-secret matching ADMIN_SECRET env var.
 */
router.post("/snapshot", async (req, res, next) => {
  try {
    const secret = req.headers["x-admin-secret"];
    if (!secret || secret !== process.env.ADMIN_SECRET) {
      return res.status(403).json({ success: false, error: "Forbidden" });
    }

    const limit = Math.min(parseInt(req.query.limit, 10) || 100, 500);

    // Compute this calendar month's top donors
    const topResult = await pool.query(
      `SELECT p.public_key, p.display_name, p.badges,
              COALESCE(SUM(d.amount_xlm), 0)::NUMERIC AS total_xlm
       FROM profiles p
       LEFT JOIN donations d
         ON p.public_key = d.donor_address
        AND d.created_at >= DATE_TRUNC('month', NOW())
        AND d.created_at <  DATE_TRUNC('month', NOW()) + INTERVAL '1 month'
       GROUP BY p.public_key, p.display_name, p.badges
       HAVING COALESCE(SUM(d.amount_xlm), 0) > 0
       ORDER BY total_xlm DESC
       LIMIT $1`,
      [limit]
    );

    if (topResult.rows.length === 0) {
      return res.json({ success: true, message: "No donations this month yet", inserted: 0 });
    }

    const monthStart = new Date();
    monthStart.setDate(1);
    monthStart.setHours(0, 0, 0, 0);
    const monthStr = monthStart.toISOString().slice(0, 10); // "YYYY-MM-01"

    const client = await pool.connect();
    try {
      await client.query("BEGIN");
      let inserted = 0;
      for (const [i, row] of topResult.rows.entries()) {
        const badge = row.badges?.[0]?.tier || null;
        await client.query(
          `INSERT INTO monthly_leaderboard
             (month, donor_address, display_name, total_xlm_that_month, badge, rank)
           VALUES ($1, $2, $3, $4, $5, $6)
           ON CONFLICT (month, donor_address)
           DO UPDATE SET
             display_name          = EXCLUDED.display_name,
             total_xlm_that_month  = EXCLUDED.total_xlm_that_month,
             badge                 = EXCLUDED.badge,
             rank                  = EXCLUDED.rank`,
          [monthStr, row.public_key, row.display_name || null, row.total_xlm, badge, i + 1]
        );
        inserted++;
      }
      await client.query("COMMIT");
      res.json({ success: true, month: monthStr, inserted });
    } catch (err) {
      await client.query("ROLLBACK");
      throw err;
    } finally {
      client.release();
    }
  } catch (e) {
    next(e);
  }
});

module.exports = router;