# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Stellar GreenPay API
  version: 1.0.0
  description: |
    OpenAPI 3.0 documentation for the backend REST API used by Stellar GreenPay.
    The API exposes project discovery, donations, profiles, leaderboard, notifications,
    ratings, updates, jobs, stats, and admin management endpoints.
  license:
    name: MIT
servers:
  - url: http://localhost:4000
    description: Local development server
  - url: https://api.stellargreenpay.com
    description: Production server
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    WalletAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Wallet JWT issued after signing a one-time Stellar challenge.
  parameters:
...
    LimitQuery:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    PeriodQuery:
      name: period
      in: query
      schema:
        type: string
        enum: [all, month, year]
        default: all
...
  responses:
    BadRequest:
      description: The request payload or query parameters are invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication is required or the provided token is invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Invalid request payload
      required:
        - error
...
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        publicKey:
          type: string
        displayName:
          type: string
          nullable: true
        totalDonatedXLM:
          type: string
        projectsSupported:
          type: integer
        topBadge:
          type: string
          nullable: true
...
    Team:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        logoUrl:
          type: string
          nullable: true
        inviteCode:
          type: string
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        memberCount:
          type: integer
        totalDonatedXLM:
          type: string
        totalCO2OffsetKg:
          type: string
        isMember:
          type: boolean
    TeamLeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        id:
          type: string
          format: uuid
        name:
          type: string
        logoUrl:
          type: string
          nullable: true
        memberCount:
          type: integer
        totalDonatedXLM:
          type: string
        totalCO2OffsetKg:
          type: string
...
paths:
...
  /api/leaderboard:
    get:
      tags: [Leaderboard]
      summary: Get the donation leaderboard
      parameters:
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/PeriodQuery'
      responses:
        '200':
          description: Leaderboard entries.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/LeaderboardEntry'
...
  /api/teams:
    post:
      tags: [Teams]
      summary: Create a team
      description: The authenticated wallet becomes the team's first member.
      security:
        - WalletAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 100
                logoUrl:
                  type: string
                  format: uri
                inviteCode:
                  type: string
                  pattern: '^[A-Za-z0-9_-]{6,20}$'
      responses:
        '201':
          description: Team created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Team'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: The wallet already belongs to a team or the invite code is taken.
...
  /api/leaderboard/teams:
    get:
      tags: [Teams, Leaderboard]
      summary: List teams ranked by combined donations
      security: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, year, all]
            default: all
      responses:
        '200':
          description: Paginated team leaderboard.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TeamLeaderboardEntry'
                  has_more:
                    type: boolean
                  next_offset:
                    type: integer
                    nullable: true
...